+++ services/orchest-webserver/client/src/Routes.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Redirect, Route, Switch, useLocation } from "react-router-dom";
 import EditJobView from "./edit-job-view/EditJobView";
 import EnvironmentEditView from "./environment-edit-view/EnvironmentEditView";
+import EnvironmentsView from "./environments-view/EnvironmentsView";
 import JobView from "./job-view/JobView";
 import JobsView from "./jobs-view/JobsView";
 import PipelineSettingsView from "./pipeline-settings-view/PipelineSettingsView";
@@ -12,7 +13,6 @@ import ProjectsView from "./projects-view/ProjectsView";
 import { getOrderedRoutes, siteMap } from "./routingConfig";
 import SettingsView from "./settings-view/SettingsView";
 import ConfigureJupyterLabView from "./views/ConfigureJupyterLabView";
-import EnvironmentsView from "./views/EnvironmentsView";
 import FilePreviewView from "./views/FilePreviewView";
 import HelpView from "./views/HelpView";
 import JupyterLabView from "./views/JupyterLabView";
