+++ services/orchest-webserver/client/src/environments-view/EnvironmentsView.tsx
@@ -1,10 +1,10 @@
-import EnvironmentList from "@/components/EnvironmentList";
 import { Layout } from "@/components/Layout";
 import ProjectBasedView from "@/components/ProjectBasedView";
 import { useCustomRoute } from "@/hooks/useCustomRoute";
 import { useSendAnalyticEvent } from "@/hooks/useSendAnalyticEvent";
 import { siteMap } from "@/routingConfig";
 import React from "react";
+import EnvironmentList from "./EnvironmentList";
 
 const EnvironmentsView: React.FC = () => {
   const { projectUuid } = useCustomRoute();
